Just Commit

+++ Home Work/Assignment4_4.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
-float harmonicSum();
+double harmonicSum();
 int main()
 {
   int n;
   scanf("%d",&n);
-  printf("%.2f",harmonicSum(n));
+  printf("%.2lf",harmonicSum(n));
 return 0;
 }
-float harmonicSum(int n)
+double harmonicSum(int n)
 {
-  float sum=0;
+  double sum=0;
   for(int i=1; i<=n; i++)
   {
     sum+=(1.0/i);
